+++ src/main/java/ir/ciph3r/mercury/utility/Utils.java
@@ -41,7 +41,7 @@ public class Utils {
 			return false;
 		}
 		try {
-			int d = Integer.parseInt(strNum);
+			double d = Double.parseDouble(strNum);
 		} catch (NumberFormatException nfe) {
 			return false;
 		}
